score-maker/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
aiohttp==3.10.9
alembic==1.13.3
asyncpg==0.29.0
python-dateutil==2.9.0.post0
greenlet==3.1.1
gunicorn
python-dotenv==1.0.1
aiokafka==0.11.0
fastapi-cache2[redis]
fastapi-utils==0.7.0
motor==3.6.0